+++ go/docker-run.sh
@@ -10,8 +10,8 @@ function mrcmd_func_go_docker_run() {
     -v "$(realpath "${GO_LIB_DIR}"):/go" \
     -v "$(realpath "${APPX_WORK_DIR}"):/opt/app" \
     --env "TZ=${APPX_TZ}" \
-    --env "APP_BIND=${GO_WEBAPP_BIND}" \
-    --env "APP_PORT=${GO_WEBAPP_PORT}" \
+    --env "APPX_SERVICE_BIND=${GO_WEBAPP_BIND}" \
+    --env "APPX_SERVICE_PORT=${GO_WEBAPP_PORT}" \
     "${GO_DOCKER_IMAGE}" \
     "$@"
 }